Rename SC_Clock.h to match conventions
